+++ services/kiloclaw/plugins/kiloclaw-morning-briefing/src/index.ts
@@ -2,12 +2,15 @@ import fs from 'node:fs/promises';
 import path from 'node:path';
 import { Type } from '@sinclair/typebox';
 import { definePluginEntry } from 'openclaw/plugin-sdk/plugin-entry';
+import { buildBriefingMarkdown, offsetDateKey, resolveBriefingPath } from './briefing-utils';
 import {
-  buildBriefingMarkdown,
-  formatBriefingMarkdownForMessage,
-  offsetDateKey,
-  resolveBriefingPath,
-} from './briefing-utils';
+  DELIVERY_CHANNELS,
+  type BriefingDeliveryResult,
+  deliverBriefingToConfiguredChannels,
+  formatDeliverySummary,
+  parseStoredDelivery,
+} from './delivery-utils';
+import { CommandExecutionError, runCommand } from './command-utils';
 import {
   filterEnabledBriefingJobs,
   pickCanonicalCronJobId,
@@ -62,25 +65,6 @@ type SourceCollectionResult = {
   summary: string;
   sectionLines: string[];
 };
-
-type DeliveryChannel = 'telegram' | 'discord' | 'slack';
-
-type BriefingDeliveryResult = {
-  channel: DeliveryChannel;
-  status: 'sent' | 'skipped' | 'failed';
-  target?: string;
-  accountId?: string;
-  reason?: 'missing_target' | 'ambiguous_target' | 'send_failed' | 'config_unavailable';
-  error?: string;
-};
-
-type DeliveryRoute = {
-  channel: DeliveryChannel;
-  target: string;
-  accountId?: string;
-};
-
-const DELIVERY_CHANNELS: DeliveryChannel[] = ['telegram', 'discord', 'slack'];
 
 function asObject(value: unknown): Record<string, unknown> {
   return typeof value === 'object' && value !== null && !Array.isArray(value)
@@ -156,30 +140,6 @@ async function writeJsonFile(filePath: string, value: unknown): Promise<void> {
   await fs.writeFile(filePath, JSON.stringify(value, null, 2), 'utf8');
 }
 
-async function runCommand(
-  api: {
-    runtime: {
-      system: {
-        runCommandWithTimeout: (
-          argv: string[],
-          options: { timeoutMs: number; cwd?: string }
-        ) => Promise<{ stdout: string; stderr: string; code: number | null }>;
-      };
-    };
-  },
-  argv: string[],
-  timeoutMs = 30_000
-): Promise<{ stdout: string; stderr: string }> {
-  const result = await api.runtime.system.runCommandWithTimeout(argv, {
-    timeoutMs,
-  });
-  if (result.code !== 0) {
-    const message = result.stderr.trim() || result.stdout.trim() || 'Command failed';
-    throw new Error(`${argv.join(' ')} failed: ${message}`);
-  }
-  return { stdout: result.stdout, stderr: result.stderr };
-}
-
 async function runCronJson(
   api: {
     runtime: {
@@ -194,11 +154,24 @@ async function runCronJson(
   argv: string[]
 ): Promise<Record<string, unknown>> {
   const [subcommand = ''] = argv;
-  const jsonUnsupported = subcommand === 'disable';
+  const jsonUnsupported = subcommand === 'disable' || subcommand === 'edit';
   const command = jsonUnsupported
     ? ['openclaw', 'cron', ...argv]
     : ['openclaw', 'cron', ...argv, '--json'];
-  const { stdout } = await runCommand(api, command, 60_000);
+  let stdout: string;
+  try {
+    ({ stdout } = await runCommand(api, command, 60_000));
+  } catch (error) {
+    if (
+      !jsonUnsupported &&
+      error instanceof CommandExecutionError &&
+      error.stderr.includes("unknown option '--json'")
+    ) {
+      ({ stdout } = await runCommand(api, ['openclaw', 'cron', ...argv], 60_000));
+    } else {
+      throw error;
+    }
+  }
   try {
     return asObject(JSON.parse(stdout));
   } catch {
@@ -374,40 +347,7 @@ async function readStoredStatus(paths: { statusPath: string }): Promise<StoredSt
     failures: Array.isArray(existing.failures)
       ? existing.failures.filter(value => typeof value === 'string')
       : [],
-    lastDelivery: Array.isArray(existing.lastDelivery)
-      ? existing.lastDelivery
-          .map(entry => asObject(entry))
-          .map(entry => {
-            const channel =
-              entry.channel === 'telegram' ||
-              entry.channel === 'discord' ||
-              entry.channel === 'slack'
-                ? entry.channel
-                : null;
-            const status =
-              entry.status === 'sent' || entry.status === 'skipped' || entry.status === 'failed'
-                ? entry.status
-                : null;
-            if (!channel || !status) {
-              return null;
-            }
-            return {
-              channel,
-              status,
-              target: typeof entry.target === 'string' ? entry.target : undefined,
-              accountId: typeof entry.accountId === 'string' ? entry.accountId : undefined,
-              reason:
-                entry.reason === 'missing_target' ||
-                entry.reason === 'ambiguous_target' ||
-                entry.reason === 'send_failed' ||
-                entry.reason === 'config_unavailable'
-                  ? entry.reason
-                  : undefined,
-              error: typeof entry.error === 'string' ? entry.error : undefined,
-            } satisfies BriefingDeliveryResult;
-          })
-          .filter((entry): entry is BriefingDeliveryResult => entry !== null)
-      : [],
+    lastDelivery: parseStoredDelivery(existing.lastDelivery),
     observedEnabled:
       typeof existing.observedEnabled === 'boolean' ? existing.observedEnabled : null,
     reconcileState:
@@ -456,312 +396,6 @@ async function patchStoredStatus(
     await writeJsonFile(paths.statusPath, next);
     return next;
   });
-}
-
-function normalizeDeliveryTarget(value: unknown): string | null {
-  if (typeof value === 'string') {
-    const trimmed = value.trim();
-    return trimmed.length > 0 ? trimmed : null;
-  }
-  if (typeof value === 'number' && Number.isFinite(value)) {
-    return String(value);
-  }
-  return null;
-}
-
-function toEnabledObjectEntries(value: unknown): Array<[string, Record<string, unknown>]> {
-  const record = asObject(value);
-  return Object.entries(record)
-    .filter((entry): entry is [string, Record<string, unknown>] => {
-      const [key, raw] = entry;
-      if (key.trim() === '' || key === '*') {
-        return false;
-      }
-      return typeof raw === 'object' && raw !== null && !Array.isArray(raw);
-    })
-    .filter(([, raw]) => raw.enabled !== false);
-}
-
-function collectTelegramFallbackTargets(rawChannelConfig: Record<string, unknown>): string[] {
-  const groups = toEnabledObjectEntries(rawChannelConfig.groups).map(([groupId]) => groupId);
-  return groups;
-}
-
-function collectDiscordFallbackTargets(rawChannelConfig: Record<string, unknown>): string[] {
-  const guildEntries = toEnabledObjectEntries(rawChannelConfig.guilds);
-  const targets = guildEntries.flatMap(([, guildConfig]) => {
-    const channels = toEnabledObjectEntries(guildConfig.channels);
-    return channels.map(([channelId]) => `channel:${channelId}`);
-  });
-  return targets;
-}
-
-function collectSlackFallbackTargets(rawChannelConfig: Record<string, unknown>): string[] {
-  const channels = toEnabledObjectEntries(rawChannelConfig.channels);
-  return channels.map(([channelId]) => `channel:${channelId}`);
-}
-
-function collectFallbackTargets(
-  channel: DeliveryChannel,
-  rawChannelConfig: Record<string, unknown>
-): string[] {
-  if (channel === 'telegram') {
-    return collectTelegramFallbackTargets(rawChannelConfig);
-  }
-  if (channel === 'discord') {
-    return collectDiscordFallbackTargets(rawChannelConfig);
-  }
-  return collectSlackFallbackTargets(rawChannelConfig);
-}
-
-function resolveDeliveryRoute(params: {
-  channel: DeliveryChannel;
-  channelsConfig: Record<string, unknown>;
-}): {
-  configured: boolean;
-  route: DeliveryRoute | null;
-  skipReason?: 'missing_target' | 'ambiguous_target';
-} {
-  const rawChannelConfig = asObject(params.channelsConfig[params.channel]);
-  if (Object.keys(rawChannelConfig).length === 0) {
-    return { configured: false, route: null };
-  }
-
-  if (rawChannelConfig.enabled === false) {
-    return { configured: false, route: null };
-  }
-
-  const accountsConfig = asObject(rawChannelConfig.accounts);
-  const defaultAccount = asObject(accountsConfig.default);
-  const defaultAccountTarget = normalizeDeliveryTarget(defaultAccount.defaultTo);
-  if (defaultAccountTarget) {
-    return {
-      configured: true,
-      route: {
-        channel: params.channel,
-        target: defaultAccountTarget,
-        accountId: 'default',
-      },
-    };
-  }
-
-  const topLevelTarget = normalizeDeliveryTarget(rawChannelConfig.defaultTo);
-  if (topLevelTarget) {
-    return {
-      configured: true,
-      route: {
-        channel: params.channel,
-        target: topLevelTarget,
-      },
-    };
-  }
-
-  const fallbackTargets = collectFallbackTargets(params.channel, rawChannelConfig);
-  if (fallbackTargets.length === 1) {
-    return {
-      configured: true,
-      route: {
-        channel: params.channel,
-        target: fallbackTargets[0],
-      },
-    };
-  }
-
-  return {
-    configured: true,
-    route: null,
-    skipReason: fallbackTargets.length > 1 ? 'ambiguous_target' : 'missing_target',
-  };
-}
-
-function readChannelsConfigFromRuntimeConfig(config: unknown): Record<string, unknown> | null {
-  const rawConfig = asObject(config);
-  if (!Object.prototype.hasOwnProperty.call(rawConfig, 'channels')) {
-    return null;
-  }
-  return asObject(rawConfig.channels);
-}
-
-async function readChannelsConfig(api: {
-  config: unknown;
-  runtime: {
-    system: {
-      runCommandWithTimeout: (
-        argv: string[],
-        options: { timeoutMs: number; cwd?: string }
-      ) => Promise<{ stdout: string; stderr: string; code: number | null }>;
-    };
-  };
-}): Promise<Record<string, unknown>> {
-  const fromRuntimeConfig = readChannelsConfigFromRuntimeConfig(api.config);
-  if (fromRuntimeConfig) {
-    return fromRuntimeConfig;
-  }
-
-  let lastError: unknown = null;
-  for (let attempt = 0; attempt < 2; attempt += 1) {
-    try {
-      const { stdout } = await runCommand(
-        api,
-        ['openclaw', 'config', 'get', 'channels', '--json'],
-        60_000
-      );
-      return asObject(JSON.parse(stdout));
-    } catch (error) {
-      const errorText = error instanceof Error ? error.message : String(error);
-      if (errorText.includes('Config path not found: channels')) {
-        return {};
-      }
-      lastError = error;
-    }
-  }
-
-  if (lastError) {
-    throw lastError;
-  }
-
-  return {};
-}
-
-function formatDeliverySummary(delivery: BriefingDeliveryResult[]): string[] {
-  if (delivery.length === 0) {
-    return ['- delivery: no configured messaging channels found'];
-  }
-
-  return delivery.map(entry => {
-    const targetSuffix = entry.target ? ` (${entry.target})` : '';
-    if (entry.status === 'sent') {
-      return `- delivery: ${entry.channel} sent${targetSuffix}`;
-    }
-    if (entry.status === 'skipped') {
-      return `- delivery: ${entry.channel} skipped (${entry.reason ?? 'unknown'})`;
-    }
-    return `- delivery: ${entry.channel} failed${targetSuffix}${entry.error ? `: ${entry.error}` : ''}`;
-  });
-}
-
-function isLikelyTimeoutError(error: unknown): boolean {
-  const message = error instanceof Error ? error.message : String(error);
-  return (
-    message.includes('operation was aborted due to timeout') ||
-    message.includes('timed out') ||
-    message.includes('ETIMEDOUT') ||
-    message.includes('Command failed')
-  );
-}
-
-async function deliverBriefingToConfiguredChannels(
-  api: {
-    config: unknown;
-    runtime: {
-      system: {
-        runCommandWithTimeout: (
-          argv: string[],
-          options: { timeoutMs: number; cwd?: string }
-        ) => Promise<{ stdout: string; stderr: string; code: number | null }>;
-      };
-    };
-    logger: { warn?: (message: string) => void };
-  },
-  markdown: string
-): Promise<BriefingDeliveryResult[]> {
-  const messageText = formatBriefingMarkdownForMessage(markdown);
-  if (!messageText) {
-    return [];
-  }
-
-  let channelsConfig: Record<string, unknown>;
-  try {
-    channelsConfig = await readChannelsConfig(api);
-  } catch (error) {
-    const errorText = error instanceof Error ? error.message : String(error);
-    api.logger.warn?.(`Morning briefing delivery config read failed: ${errorText}`);
-    return DELIVERY_CHANNELS.map(channel => ({
-      channel,
-      status: 'failed',
-      reason: 'config_unavailable',
-      error: errorText,
-    }));
-  }
-
-  const delivery: BriefingDeliveryResult[] = [];
-  const routes: DeliveryRoute[] = [];
-
-  for (const channel of DELIVERY_CHANNELS) {
-    const resolution = resolveDeliveryRoute({ channel, channelsConfig });
-    if (!resolution.configured) {
-      continue;
-    }
-    if (!resolution.route) {
-      delivery.push({
-        channel,
-        status: 'skipped',
-        reason: resolution.skipReason ?? 'missing_target',
-      });
-      continue;
-    }
-    routes.push(resolution.route);
-  }
-
-  for (const route of routes) {
-    const argv = [
-      'openclaw',
-      'message',
-      'send',
-      '--channel',
-      route.channel,
-      '--target',
-      route.target,
-      '--message',
-      messageText,
-    ];
-    if (route.accountId) {
-      argv.push('--account', route.accountId);
-    }
-    try {
-      await runCommand(api, argv, 120_000);
-      delivery.push({
-        channel: route.channel,
-        status: 'sent',
-        target: route.target,
-        accountId: route.accountId,
-      });
-    } catch (error) {
-      if (isLikelyTimeoutError(error)) {
-        try {
-          await runCommand(api, argv, 120_000);
-          delivery.push({
-            channel: route.channel,
-            status: 'sent',
-            target: route.target,
-            accountId: route.accountId,
-          });
-          continue;
-        } catch (retryError) {
-          delivery.push({
-            channel: route.channel,
-            status: 'failed',
-            reason: 'send_failed',
-            target: route.target,
-            accountId: route.accountId,
-            error: retryError instanceof Error ? retryError.message : String(retryError),
-          });
-          continue;
-        }
-      }
-
-      delivery.push({
-        channel: route.channel,
-        status: 'failed',
-        reason: 'send_failed',
-        target: route.target,
-        accountId: route.accountId,
-        error: error instanceof Error ? error.message : String(error),
-      });
-    }
-  }
-
-  return delivery;
 }
 
 async function ensureCronJob(
